griffin/refactor-add-objects-page-path
