+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/OpsPage/OpPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {SimplePageLayout} from './common/SimplePageLayout';
+import {SimplePageLayout} from '../common/SimplePageLayout';
 import {FilterableOpVersionsTable} from './OpVersionsPage';
 
 export const OpPage: React.FC<{
