+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ScorersPage/CoreScorersTab.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {FilterableObjectVersionsTable} from '../ObjectVersionsPage';
+import {FilterableObjectVersionsTable} from '../ObjectsPage/ObjectVersionsTable';
 
 export const ProgrammaticScorersTab: React.FC<{
   entity: string;
